Dividing statistics Tools into multiple partial files.
